Moves exit paths logic to separate helper class
